+++ src/main/java/de/pmcp/hungergames/CMDS/TabCompletion.java
@@ -17,9 +17,9 @@ public class TabCompletion implements TabCompleter {
     public List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String[] args) {
         if (command.getName().equals("timer")) {
             if (args.length == 1)
-                return Arrays.asList((DayTimer.timerPaused ? "run" : "pause"), "set", "status");
+                return Arrays.asList((DayTimer.timerPaused ? "resume" : "pause"), "set", "status");
             else if (args[0].equals("set") && args.length == 2)
-                return Arrays.asList("-600", "-60", "0", "600", "1200", "1800", "2400");
+                return Arrays.asList("1200", "2400", "3600", "4800", "6000", "7200", "8400");
 
         }
         return Arrays.asList();
