+++ src/main/java/de/pmcp/hungergames/game/DailyNews.java
@@ -1,0 +1,4 @@
+package de.pmcp.hungergames.game;
+
+public class DailyNews {
+}
