+++ src/main/java/de/pmcp/hungergames/game/DayTimer.java
@@ -1,13 +1,10 @@
 package de.pmcp.hungergames.game;
 
 import de.pmcp.hungergames.CMDS.Freeze;
-import de.pmcp.hungergames.CMDS.Timer;
 import de.pmcp.hungergames.game.Engine;
 import de.pmcp.hungergames.main;
 import org.apache.commons.lang.ArrayUtils;
 import org.bukkit.Bukkit;
-import org.bukkit.Sound;
-import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitScheduler;
 
 public class DayTimer {
@@ -22,46 +19,36 @@ public class DayTimer {
         int day = Engine.day;
         scheduler.runTaskTimer(main.plugin,task -> {
             if (timerPaused) return;
-            //Start des Countdowns
+
             if (time[0] < 0) {
-                int dist = time[0] * -1; //Abstand zum start des Tages
+                int dist = time[0] * -1;
                 if (ArrayUtils.contains(timerPoints, dist))
                     if (day == 0)
                         Bukkit.broadcastMessage("§e[§6Hungergames§e] §4Der Tag beginnt in §c" + (dist <= -60 ? dist / 60 + " §4Minuten" : dist + " §4Sekunden"));
                     else
                         Bukkit.broadcastMessage("§e[§6Hungergames§e] §4Die Spiele starten in §c" + (dist <= -60 ? dist / 60 + " §4Minuten" : dist + " §4Sekunden"));
-                if (dist == 5 && day > 1) Info.DailyNews();
             }
-            //Start des Tages
             else if (time[0] == 0)  {
-                Engine.day ++;
-                Freeze.isfreeze = false;
-                for (Player player : Bukkit.getOnlinePlayers()) player.playSound(player, Sound.ENTITY_GOAT_SCREAMING_AMBIENT, 10F, 0.8F);
-
-                if (day == 1) {
+                if (day == 0)
                     Bukkit.broadcastMessage("§6§lDie Spiele sind gestartet: \n" + "§4§lViel Glück!");
-                }
-                else {
+                else
                     Bukkit.broadcastMessage("§6§lDer Tag ist gestartet: \n" + "§4§lViel Glück!");
-                }
+                Freeze.isfreeze = false;
             }
-            //Während dem Tag
             else if (time[0] < sessionLength) {
-                int dist = sessionLength - time[0]; //Abstand zum Ende des Tages
+                int dist = sessionLength - time[0];
                 if (ArrayUtils.contains(timerPoints, dist))
                     if (day == 7)
                         Bukkit.broadcastMessage("§e[§6Hungergames§e] §4Die Spiele laufen noch §c" + (dist >= 60 ? dist / 60 + " §4Minuten" : dist + " §4Sekunden"));
                     else
                         Bukkit.broadcastMessage("§e[§6Hungergames§e] §4Der Tag endet in §c" + (dist >= 60 ? dist / 60 + " §4Minuten" : dist + " §4Sekunden"));
             }
-            //Ende des Tages
             else {
                 if (day == 7)
                     Bukkit.broadcastMessage("§6§lDie Spiele sind geendet!");
                 else
                     Bukkit.broadcastMessage("§6§lDer Tag ist geendet!");
-                Info.DailyNews();
-                Freeze.isfreeze = true;
+                Freeze.isfreeze = false;
                 timerPaused = true;
                 time[0] = -11;
             }
