+++ src/main/java/de/pmcp/hungergames/main.java
@@ -3,12 +3,11 @@ package de.pmcp.hungergames;
 //Hungergames Dateien
 import de.pmcp.hungergames.CMDS.Adminmsg;
 import de.pmcp.hungergames.CMDS.Timer;
-import de.pmcp.hungergames.game.Info;
+import de.pmcp.hungergames.game.InfoBar;
 import de.pmcp.hungergames.game.Death;
 import de.pmcp.hungergames.game.DayTimer;
 import de.pmcp.hungergames.CMDS.TabCompletion;
 import de.pmcp.hungergames.CMDS.Freeze;
-import de.pmcp.hungergames.game.Volcano;
 import de.pmcp.hungergames.tools.Freezer;
 
 import org.bukkit.Bukkit;
@@ -33,12 +32,11 @@ public final class main extends JavaPlugin {
 
         //listener register
         PluginManager pluginManager = Bukkit.getPluginManager();
-        //pluginManager.registerEvents(new Volcano(), this);
+        //pluginManager.registerEvents(new volcano(), this);
         pluginManager.registerEvents(new Freezer(), this);
         Freezer.effects();
         pluginManager.registerEvents(new Death(), this);
-        Info.info();
-        pluginManager.registerEvents(new Info(), this);
+        InfoBar.info();
         DayTimer daytimer = new DayTimer(); daytimer.timer();
     }
