--- src/main/java/de/pmcp/hungergames/tools/Database.java
@@ -1,14 +1,0 @@
-package de.pmcp.hungergames.tools;
-
-public class Database {
-    //Liest daten aus dem Speicher/Config
-    public static String[] get() {
-        String[] data = {};
-
-        return data;
-    }
-    //Schreibt Daten in den Speicher/Config
-    public static void write(String[] data) {
-
-    }
-}
