+++ src/main/resources/plugin.yml
@@ -44,7 +44,7 @@ commands:
     permission: pmcp.admin
     permission-message: Nicht jeder Mensch.
     description: Timer für das Spiel (einen tag)
-    usage: §7/timer §f[pause, run, set <sekunden>, status]
+    usage: §7/timer §f[pause, resume, set <sekunden>, status]
 
 
 #Kommentar für mich: listen angeben mit [..., ...] oder ... oder -... jede zeile
